feat(headless): Bayer debayer, calibration pipeline, software binning
- FITS/MJPEG colour pipeline: BAYERPAT-aware bilinear demosaic at full
  resolution for stills, in-process 2×2 binning debayer for the streaming
  MJPEG (PlayerOne CCD_VIDEO_FORMAT exposes only POA_RAW8/16). Mono cameras
  pass through unchanged.
- Hardware binning is now forced to 1×1 for colour sensors so the Bayer
  pattern survives capture; the user's requested binning is applied as a
  server-side average pool after debayer (preserves colour at any binning).
- INDI capture timeout raised from exposure+30s to exposure+90s — large CMOS
  readout could push the previous margin into 504s on otherwise healthy 30s
  exposures.
- Hot-path: cached `_hardwareBinned` + 30-frame recheck in
  CameraStreamService.OnBlobReceived (was calling TryBuildCameraInfo every
  BLOB). Dropped dead `_streamGain` field.
- New calibration / flat-wizard / live-stack / star-detection services and
  capture/calibration library stores.
- Switch / weather controllers, INDI client refactor (Indi/ folder).

+++ NINA.Headless/Models/MoveRequest.cs
@@ -3,6 +3,6 @@ namespace NINA.Headless.Models;
 public record MoveRequest
 {
     public string Direction { get; init; } = "N";
-    public int Rate { get; init; } = 1;
+    public double Rate { get; init; } = 1.0;
     public int Duration { get; init; } = 500;
 }
